refactor(users mo)[OK-103] Move users dir from auth to modules
