+++ backend/src/modules/auth/auth.controller.ts
@@ -12,7 +12,7 @@ import { AuthService } from './auth.service';
 import { LocalAuthGuard } from './guards/local.auth.guard';
 import { JwtAuthGuard } from './guards/jwt.auth.guard';
 import { SafeUserDoc } from '../../shared/interfaces/SafeUserDoc';
-import { CreateUserDto } from './users/users.dto';
+import { CreateUserDto } from '../users/users.dto';
 import { GoogleAuthGuard } from './guards/google.auth.guard';
 import { Request, Response } from 'express';
 import { AuthToken } from '../../shared/interfaces/AuthToken';
