+++ backend/src/modules/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
 import { MongooseModule } from '@nestjs/mongoose';
-import { User, UserSchema } from '../../../schemas/user.schema';
+import { User, UserSchema } from '../../schemas/user.schema';
 
 @Module({
   imports: [
